tools/aseprite/requirements.txt
PyYAML>=6.0
Pillow>=10.0
numpy>=1.26
# Phase 1.5: AI background removal (U²-Net via rembg)
# onnxruntime-gpu has no aarch64 wheel as of 2026-04; CPU is fine for 144-image batches (~8 min total).
rembg>=2.0.75
onnxruntime>=1.16.0
